+++ src/utils/core/citation_utils.tsx
@@ -1,5 +1,5 @@
-import { escapeString, removeBraces, removeInlineCodeBlocks } from "@/utils/string_processing/string_utils";
-import { equationBlockBracePattern, inlineMathPattern, isCodeBlockToggle, matchCitationsInLine, matchNestedCitation } from "@/utils/string_processing/regexp_utils";
+import { removeBraces, removeInlineCodeBlocks } from "@/utils/string_processing/string_utils";
+import { isCodeBlockToggle, matchCitationsInLine } from "@/utils/string_processing/regexp_utils";
 import { extractLastNumberFromTag, extractPrefixBeforeLastNumber } from "@/utils/parsers/equation_parser";
 
 export interface CitationRef {
@@ -273,237 +273,6 @@ export function parseCitationsInMarkdown(md: string): CitationRef[] {
     }
     return result;
 }
-
-////////////////////  Following Functions are for PDF export usage  ///////////////////////// 
-
-export interface SpanStyles {
-    citationColorInPdf: string;
-}
-
-/**
- * Replaces inline citations in markdown with <span> tags
- * Ignores: 1. Multiline code blocks 2. Inline code blocks 3. Citations in display math ($$ $$)
- * 
- * @param markdown - Original markdown string
- * @param rangeSymbol - Range symbol, defaults to '~', if null, not use continuous equation tags 
- * @param validDelimiters - Valid delimiters, defaults to ['.', '-']
- * @param fileDelimiter - File delimiter, defaults to '^'
- * @returns Processed markdown string
- */
-export function replaceCitationsInMarkdownWithSpan(
-    markdown: string,
-    prefix: string,
-    rangeSymbol: string | null,
-    validDelimiters: string[],
-    fileDelimiter: string,
-    multiCitationDelimiter = ',',
-    citationFormat = '(#)',
-    spanStyles = {} as SpanStyles
-): string {
-    if (!markdown.trim()) return markdown;
-
-    const lines = markdown.split('\n');
-    let inMultilineCodeBlock = false;
-    let inDisplayMath = false;
-
-    const processedLines = lines.map((line, lineNum) => {
-        let processedLine = line;
-        // Handles multiline code block state
-        if (isCodeBlockToggle(line)) {
-            inMultilineCodeBlock = !inMultilineCodeBlock;
-        }
-        if (inMultilineCodeBlock) {
-            return line; // In code block - skip processing
-        }
-        // Handles display math block state
-        const displayMathMatches = line.match(equationBlockBracePattern);
-        if (displayMathMatches && displayMathMatches.length % 2 !== 0) inDisplayMath = !inDisplayMath;
-        if (inDisplayMath) {
-            return line; // In display math block - skip processing
-        }
-        processedLine = processInlineReferences(
-            line, prefix, rangeSymbol, validDelimiters, fileDelimiter, multiCitationDelimiter, citationFormat, spanStyles
-        );
-        return processedLine;
-    });
-    return processedLines.join('\n');
-}
-
-/**
- * Processes inline citations to span while avoiding inline code blocks
- */
-function processInlineReferences(
-    line: string,
-    prefix: string,
-    rangeSymbol: string | null,  // if null, not use continuous equation tags 
-    validDelimiters: string[],
-    fileDelimiter: string, // if null, not use file citations 
-    multiCitationDelimiter = ',',
-    citationFormat = '(#)',
-    spanStyles = {} as SpanStyles
-): string {
-    // First marks all inline code block positions
-    const codeBlockRanges: Array<{ start: number, end: number }> = [];
-    let i = 0;
-
-    while (i < line.length) {
-        if (line[i] === '`' && (i === 0 || line[i - 1] !== '\\')) {
-            const start = i;
-            i++; // Skips opening backtick 
-            // Finds closing backtick 
-            while (i < line.length) {
-                if (line[i] === '`' && line[i - 1] !== '\\') {
-                    codeBlockRanges.push({ start, end: i });
-                    break;
-                }
-                i++;
-            }
-        }
-        i++;
-    }
-
-    // Checks if position is within a code block
-    const isInCodeBlock = (pos: number): boolean => {
-        return codeBlockRanges.some(range => pos >= range.start && pos <= range.end);
-    };
-    // Find all potential inline math expressions first
-    const dollarPositions: Array<{ pos: number, type: 'single' | 'double' }> = [];
-    // Single-pass scan: track consecutive backslashes so we don't backward-scan per '$'
-    i = 0;
-    let backslashRun = 0; // count of consecutive backslashes immediately before current index
-    while (i < line.length) {
-        const ch = line[i];
-        if (ch === '\\') {
-            backslashRun++;
-            i++;
-            continue;
-        }
-        if (ch === '$') {
-            const isEscaped = (backslashRun % 2 === 1);
-            if (!isEscaped) {
-                if (i + 1 < line.length && line[i + 1] === '$') {
-                    dollarPositions.push({ pos: i, type: 'double' });
-                    i += 2; // skip the second '$'
-                } else {
-                    dollarPositions.push({ pos: i, type: 'single' });
-                    i += 1;
-                }
-                backslashRun = 0;
-                continue;
-            }
-        }
-        // Any non-backslash, non-handled character resets the run
-        backslashRun = 0;
-        i++;
-    }
-
-    // Finds all valid inline math patterns by dollar positions  
-    const inlineMathRanges: Array<{ start: number, end: number }> = [];
-    for (let j = 0; j < dollarPositions.length - 1; j++) {
-        const current = dollarPositions[j];
-        const next = dollarPositions[j + 1];
-        if (current.type === 'single' && next.type === 'single') {
-            // Check if both positions are outside code blocks
-            if (isInCodeBlock(current.pos) || isInCodeBlock(next.pos)) continue;
-            // match the math pattern 
-            const content = line.substring(current.pos, next.pos + 1); 
-            if (new RegExp(inlineMathPattern).exec(content)) {
-                // only add if it's a inline math pattern  
-                inlineMathRanges.push({ start: current.pos, end: next.pos });
-                j++; // Skip the next position as it's already paired
-            }
-        }
-    }
-
-    // Find citations within valid inline math ranges
-    const matches: Array<{
-        mathStart: number,
-        mathEnd: number,
-        content: string,
-        citations: string[]
-    }> = [];
-    for (const mathRange of inlineMathRanges) {
-        // this has been a valid math content checked
-        const mathContent = line.substring(mathRange.start + 1, mathRange.end); // +1 to skip opening $
-        const match = matchNestedCitation(mathContent, prefix);  
-        // Check for exactly one \ref{} in this math expression
-        
-        if (match) {
-            const citations = match.label.split(multiCitationDelimiter).map(c => c.trim()).filter(c => c.length > 0);
-            matches.push({
-                mathStart: mathRange.start,
-                mathEnd: mathRange.end + 1, // +1 to include closing $
-                content: mathContent,
-                citations
-            });
-        }
-    }
-
-    // Replace from end to start 
-    let result = line;
-    matches.toReversed().forEach(({ mathStart, mathEnd, citations }) => {
-        // Combines continuous citation tags
-        const combinedCitations = (rangeSymbol === null) ?
-            citations :
-            combineContinuousCitationTags(
-                citations,
-                rangeSymbol,
-                validDelimiters,
-                fileDelimiter
-            );
-
-        // Generates replacement HTML
-        const replacement = generateCitationSpans(
-            combinedCitations, fileDelimiter, multiCitationDelimiter, citationFormat, spanStyles
-        );
-        // Performs replacement
-        result = result.substring(0, mathStart) + replacement + result.substring(mathEnd);
-    });
-    return result;
-}
-
-
-/**
- * Default styles for PDF print citations
- */
-const DEFAULT_CONTAINER_STYLE = 'cursor: default;';
-const DEFAULT_CITATION_STYLE = 'text-decoration: none; cursor: pointer;';
-
-/**
- * Generates span tags for citations
- */
-export function generateCitationSpans(
-    citations: string[],
-    fileDelimiter: string,
-    multiCitationDelimiter = ',',
-    citationFormat = '(#)',
-    spanStyles = {} as SpanStyles
-): string {
-    const spans = citations.map((citation, index) => {
-        const { local, crossFile } = splitFileCitation(citation, fileDelimiter);
-        const citationColorInPdf = spanStyles.citationColorInPdf || '#000000';
-
-        // Combine default styles with color
-        const containerStyle = escapeString(
-            DEFAULT_CONTAINER_STYLE + ` color: ${citationColorInPdf};`,
-            "\""
-        );
-
-        const citationStyle = escapeString(
-            DEFAULT_CITATION_STYLE + ` color: ${citationColorInPdf};`,
-            "\""
-        );
-        let result = `<span style="${containerStyle}">` + citationFormat.replace('#', `<span style="${citationStyle}">${local}</span>`);
-        result += '</span>';
-        if (crossFile) {
-            result += `${'[^' + crossFile + ']'}`;
-        } 
-
-        return result;
-    });
-    return spans.join(multiCitationDelimiter + ' '); // Joins all spans with comma separator
-}
-
 
 ///////////////////////// Auto Complete Functions  //////////////////// 
 export function extractAutoCompleteInputTag(content: string, delimiter: string): string {
